+++ kamehouse-shell/bin/kamehouse/install-kamehouse-shell.sh
@@ -275,6 +275,8 @@ logKameHouseShellStatus() {
   log.info "Deployed kamehouse-shell status"
   log.info "ls -lh ${COL_CYAN_STD}${KAMEHOUSE_SHELL_PATH}"
   ls -lh "${KAMEHOUSE_SHELL_PATH}"
+  log.info "shell-version.txt"
+  cat "${KAMEHOUSE_SHELL_PATH}/conf/shell-version.txt"
 }
 
 log.info() {